Move reading functions back to `ReplicationBuffer`
I think we should either move all readings there or nothing. But I don't
think that reading should be a part of `ReplicationBuffer`, I would prefer
to see them as free functions or as a part of some reader-like struct.
Also saves us from `Vec<(Entity, Entity)>` allocation.
